+++ node/proxy/api/src/models.ts
@@ -1,3 +1,11 @@
 export interface ValidationResult {
   valid: boolean
 }
+
+export const services = ['bebop', 'chainflip', 'mayachain', 'nearintents', 'portals', 'thorchain', 'zrx'] as const
+export type Service = (typeof services)[number]
+
+export interface AffiliateRevenueResponse {
+  byService: Record<Service, number>
+  totalUsd: number
+}
