+++ node/proxy/api/src/swagger.json
@@ -67,6 +67,66 @@
 				],
 				"type": "object",
 				"additionalProperties": false
+			},
+			"Record_Service.number_": {
+				"properties": {
+					"bebop": {
+						"type": "number",
+						"format": "double"
+					},
+					"chainflip": {
+						"type": "number",
+						"format": "double"
+					},
+					"mayachain": {
+						"type": "number",
+						"format": "double"
+					},
+					"nearintents": {
+						"type": "number",
+						"format": "double"
+					},
+					"portals": {
+						"type": "number",
+						"format": "double"
+					},
+					"thorchain": {
+						"type": "number",
+						"format": "double"
+					},
+					"zrx": {
+						"type": "number",
+						"format": "double"
+					}
+				},
+				"required": [
+					"bebop",
+					"chainflip",
+					"mayachain",
+					"nearintents",
+					"portals",
+					"thorchain",
+					"zrx"
+				],
+				"type": "object",
+				"description": "Construct a type with a set of properties K of type T"
+			},
+			"AffiliateRevenueResponse": {
+				"properties": {
+					"byService": {
+						"$ref": "#/components/schemas/Record_Service.number_"
+					},
+					"totalUsd": {
+						"type": "number",
+						"format": "double"
+					}
+				},
+				"required": [
+					"byService",
+					"totalUsd"
+				],
+				"type": "object",
+				"additionalProperties": false
 			}
 		},
 		"securitySchemes": {}
@@ -85,7 +145,7 @@
 				"operationId": "ValidateAddress",
 				"responses": {
 					"200": {
-						"description": "transaction payload",
+						"description": "validation result",
 						"content": {
 							"application/json": {
 								"schema": {
@@ -145,6 +205,80 @@
 						"required": true,
 						"schema": {
 							"type": "string"
+						}
+					}
+				]
+			}
+		},
+		"/api/v1/affiliate/revenue": {
+			"get": {
+				"operationId": "GetAffilitateRevenue",
+				"responses": {
+					"200": {
+						"description": "affiliate revenue",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/AffiliateRevenueResponse"
+								}
+							}
+						}
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/BadRequestError"
+								}
+							}
+						}
+					},
+					"422": {
+						"description": "Validation Error",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ValidationError"
+								}
+							}
+						}
+					},
+					"500": {
+						"description": "Internal Server Error",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/InternalServerError"
+								}
+							}
+						}
+					}
+				},
+				"description": "Get affiliate revenue",
+				"tags": [
+					"Affiliate Revenue"
+				],
+				"security": [],
+				"parameters": [
+					{
+						"description": "start timestamp (unix seconds)",
+						"in": "query",
+						"name": "startTimestamp",
+						"required": true,
+						"schema": {
+							"format": "double",
+							"type": "number"
+						}
+					},
+					{
+						"description": "end timestamp (unix seconds)",
+						"in": "query",
+						"name": "endTimestamp",
+						"required": true,
+						"schema": {
+							"format": "double",
+							"type": "number"
 						}
 					}
 				]
