+++ node/proxy/sample.env
@@ -5,3 +5,4 @@ COINGECKO_API_KEY=
 ZERION_API_KEY=
 ZRX_API_KEY=
 PORTALS_API_KEY=
+BEBOP_API_KEY=
